Harden judge-eval resume compaction and logprob precision
Compact the judge resume JSONL to one record per comparison before
re-attempting failed judgments. Previously each resume re-appended a new
line for every non-ok comparison while leaving the stale one in place, so
repeated resumes of a flaky remote judge grew the file without bound
(load-time dedup kept results correct but the file bloated). main() now
rewrites the file atomically to keep only the ok records, fsyncing resumed
work before any new attempt, so non-resume runs and crash-safety are
unchanged.

Upcast logits to float32 before log_softmax in
sequence_logp_sums_and_counts, matching sequence_logps_with_token_kl. This
is a no-op under today's fp32 eval loading but keeps summed log-probs (and
every margin/reward derived from them) accurate if eval models are ever
loaded in bf16/fp16.

Add a regression test covering resume compaction of a failed-then-ok judge.

+++ src/cpo_trl/evaluation/teacher_forced.py
@@ -326,7 +326,11 @@ def sequence_logp_sums_and_counts(
     outputs = model(input_ids=input_ids, attention_mask=attention_mask)
     logits = outputs.logits[:, :-1, :]
     labels = input_ids[:, 1:]
-    token_logps = torch.gather(torch.log_softmax(logits, dim=-1), 2, labels.unsqueeze(-1)).squeeze(-1)
+    # Upcast to float32 before log_softmax so summed log-probs stay accurate when
+    # the model runs in a reduced precision (matches sequence_logps_with_token_kl).
+    token_logps = torch.gather(
+        torch.log_softmax(logits.float(), dim=-1), 2, labels.unsqueeze(-1)
+    ).squeeze(-1)
     token_mask = attention_mask[:, 1:].bool()
     positions = torch.arange(labels.shape[1], device=labels.device).unsqueeze(0)
     response_mask = positions >= (starts.to(labels.device).unsqueeze(1) - 1)
